docker/requirements.txt
boto3
requests
python-dotenv
urllib3
